# swagger.yaml
swagger: "2.0"
info:
  description: >
    This is the API documentation of a RESTful web service, implemented
    in order to sustain the Software Engineering II exam at University of Trento.

    #####Description

    This API enables teachers to test the knowledge of their students by assigning
    them some tasks to do. There will also be some features in order to eventually
    perform a peer review session on some task
  version: "1.0.0"
  title: "EsAPI"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "giovanni.guadagnini@studenti.unitn.it"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "teamrocketproject.herokuapp.com"
basePath: "/"
...
schemes:
- "https"
paths:
...
  /users:
    post:
      tags:
      - "User"
      summary: "Register as a new user"
      description: "One time thing in order to register itself into the application and use it"
      operationId: "addUser"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "User object that represents the profile newly registered user"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/User"
        405:
          description: "Invalid input"
    get:
      tags:
      - "User"
      summary: "Return the list of users"
      description: "Every user can see a complete list of users that are related to them in some ways (teachers and/or students)"
      operationId: "getUsers"
      produces:
      - "application/json"
      parameters:
      - in: query
        name: "enrolledAfter"
        type: integer
        description: "Filter students enrolled after a certain year"
        default: 1990
      - in: query
        name: "enrolledBefore"
        type: integer
        description: "Filter students enrolled before a certain year (default current year)"
      responses:
        200:
            description: "Successful operation"
            schema:
              type: "array"
              items:
                $ref: "#/definitions/User"
  /users/{id}:
    parameters:
    - in: path
      name: id
      required: true
      type: integer
      description: The user ID
    get:
      tags:
      - "User"
      summary: "Find user by ID"
      description: "Could be performed by the user itself which wants to see its own profile, but also from other users that has the right to inspect other profile's data"
      operationId: "getUserById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of user to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid ID supplied"
        404:
          description: "User not found"
    put:
      tags:
      - "User"
      summary: "Update user informations"
      description: "Performed by the user itself which wants to update its own profile"
      operationId: "updateUserByID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "User object that needs to be updated"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid ID supplied"
        404:
          description: "User not found"
    delete:
      tags:
      - "User"
      summary: "Delete user with given ID"
      description: "Performed by the user itself which wants to delete its own profile"
      operationId: "deleteUser"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of user to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid ID supplied"
        404:
          description: "User not found"
...
definitions:
...
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        description: "Unique identifier for any user of the system"
      name:
        type: "string"
        example: "Bob"
        description: "Name of the user"
      password:
        type: "string"
        description: "Password of the user"
      surname:
        type: "string"
        example: "Strongman"
        description: "Surname of the user"
      email:
        type: "string"
        format: "bob.strongman@studenti.unitn.it"
        description: "Email of the user"
      exam_eval:
        type: "array"
        items:
          $ref: "#/definitions/ExamEval"
        description: "Array that will expose the final marks of every exam taken by the user"
      born:
        $ref: "#/definitions/Datetime"
        description: "Birth date of the user"
      enrolment:
        $ref: "#/definitions/Datetime"
        description: "Year of enrolment"
      submissions:
        type: "array"
        items:
          $ref: "#/definitions/Submission"
        description: "Array that will expose all the submissions object that has been delivered during any exam"
...
  Submission:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        description: "Unique identifier for a task submission"
      task_type:
        enum: ["open", "single_c", "multiple_c", "submit"]
        description: "Type of task submission"
      question:
        type: "object"
        description: "Question formatted in a proper way: this is defined based on the task_type (update on this topic very soon)"
      answer:
        type: "string"
        description: "Answer to a task submission formatted in a proper way: this is defined based on the task_type (update on this topic very soon)"
      id_user:
        type: "integer"
        format: "int64"
        description: "Unique identifier for the user that has performed the submission"
      id_exam:
        type: "integer"
        format: "int64"
        description: "Unique identifier for the exam in which the submission fit into"
      completed:
        type: "boolean"
        description: "Mark the submission has finished (it means you can't change the answer anymore)"
      comment_peer:
        type: "array"
        items:
          type: "string"
        description: "Array of comments delivered during a peer review for a task submission"
      comment:
        type: "string"
        description: "Teacher comment for a task submission"
      points:
        type: "integer"
        format: "int32"
        description: "Max point that could be assigned for a complete and satisfactory answer for this task (helps to build a final mark for the exam)"
      earned_points:
        type: "integer"
        format: "int32"
        description: "Teacher evaluation for a task submission (it has be less than the max points)"
  Datetime:
    type: "object"
    properties:
      year:
        type: "integer"
        format: "int64"
        minimum: 1900
      month:
        type: "integer"
        format: "int32"
        minimum: 1
        maximum: 12
      day:
        type: "integer"
        format: "int32"
        minimum: 1
        maximum: 31
      hour:
        type: "integer"
        format: "int32"
        minimum: 0
        maximum: 23
      minute:
        type: "integer"
        format: "int32"
        minimum: 0
        maximum: 60
      second:
        type: "integer"
        format: "int32"
        minimum: 0
        maximum: 60
  ExamEval:
    type: "object"
    properties:
      id_exam:
        type: "integer"
        format: "int64"
        description: "ID of the exam the mark is related to"
      mark:
        type: "integer"
        format: "int32"
        minimum: 0
        description: "Final mark for a specific exam"
        maximum: 31